third_party/blink/web_tests/fast/dom/dom-constructors-expected.txt

This test checks that all but a handful of dom constructors throw exceptions, and the rest return reasonable objects. It also tests that those constructors have higher precedence than a document element with the same ID or name.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS TryAllocate('Attr') is 'exception'
PASS TryAllocate('CharacterData') is 'exception'
PASS TryAllocate('CDATASection') is 'exception'
PASS TryAllocate('DocumentType') is 'exception'
PASS TryAllocate('Element') is 'exception'
PASS TryAllocate('HTMLDocument') is 'exception'
PASS TryAllocate('Node') is 'exception'
PASS TryAllocate('ProcessingInstruction') is 'exception'
PASS TryAllocate('HTMLAllCollection') is 'exception'
PASS TryAllocate('HTMLAnchorElement') is 'exception'
PASS TryAllocate('HTMLAreaElement') is 'exception'
PASS TryAllocate('HTMLBaseElement') is 'exception'
PASS TryAllocate('HTMLBodyElement') is 'exception'
PASS TryAllocate('HTMLBRElement') is 'exception'
PASS TryAllocate('HTMLButtonElement') is 'exception'
PASS TryAllocate('HTMLCanvasElement') is 'exception'
PASS TryAllocate('HTMLDirectoryElement') is 'exception'
PASS TryAllocate('HTMLDivElement') is 'exception'
PASS TryAllocate('HTMLDListElement') is 'exception'
PASS TryAllocate('HTMLEmbedElement') is 'exception'
PASS TryAllocate('HTMLFieldSetElement') is 'exception'
PASS TryAllocate('HTMLFontElement') is 'exception'
PASS TryAllocate('HTMLFormElement') is 'exception'
PASS TryAllocate('HTMLFrameElement') is 'exception'
PASS TryAllocate('HTMLFrameSetElement') is 'exception'
PASS TryAllocate('HTMLHeadingElement') is 'exception'
PASS TryAllocate('HTMLHeadElement') is 'exception'
PASS TryAllocate('HTMLHRElement') is 'exception'
PASS TryAllocate('HTMLHtmlElement') is 'exception'
PASS TryAllocate('HTMLIFrameElement') is 'exception'
PASS TryAllocate('HTMLImageElement') is 'exception'
PASS TryAllocate('HTMLInputElement') is 'exception'
PASS TryAllocate('HTMLLabelElement') is 'exception'
PASS TryAllocate('HTMLLegendElement') is 'exception'
PASS TryAllocate('HTMLLIElement') is 'exception'
PASS TryAllocate('HTMLLinkElement') is 'exception'
PASS TryAllocate('HTMLMapElement') is 'exception'
PASS TryAllocate('HTMLMarqueeElement') is 'exception'
PASS TryAllocate('HTMLMenuElement') is 'exception'
PASS TryAllocate('HTMLMetaElement') is 'exception'
PASS TryAllocate('HTMLModElement') is 'exception'
PASS TryAllocate('HTMLObjectElement') is 'exception'
PASS TryAllocate('HTMLOListElement') is 'exception'
PASS TryAllocate('HTMLOptGroupElement') is 'exception'
PASS TryAllocate('HTMLOptionElement') is 'exception'
PASS TryAllocate('HTMLParagraphElement') is 'exception'
PASS TryAllocate('HTMLParamElement') is 'exception'
PASS TryAllocate('HTMLPreElement') is 'exception'
PASS TryAllocate('HTMLQuoteElement') is 'exception'
PASS TryAllocate('HTMLScriptElement') is 'exception'
PASS TryAllocate('HTMLSelectElement') is 'exception'
PASS TryAllocate('HTMLStyleElement') is 'exception'
PASS TryAllocate('HTMLTableCaptionElement') is 'exception'
PASS TryAllocate('HTMLTableColElement') is 'exception'
PASS TryAllocate('HTMLTableElement') is 'exception'
PASS TryAllocate('HTMLTableSectionElement') is 'exception'
PASS TryAllocate('HTMLTableCellElement') is 'exception'
PASS TryAllocate('HTMLTableRowElement') is 'exception'
PASS TryAllocate('HTMLTextAreaElement') is 'exception'
PASS TryAllocate('HTMLTitleElement') is 'exception'
PASS TryAllocate('HTMLUListElement') is 'exception'
PASS TryAllocate('HTMLElement') is 'exception'
PASS TryAllocate('CanvasRenderingContext2D') is 'exception'
PASS TryAllocate('CSSFontFaceRule') is 'exception'
PASS TryAllocate('CSSImportRule') is 'exception'
PASS TryAllocate('CSSMediaRule') is 'exception'
PASS TryAllocate('CSSPageRule') is 'exception'
PASS TryAllocate('CSSRule') is 'exception'
PASS TryAllocate('CSSRuleList') is 'exception'
PASS TryAllocate('CSSStyleDeclaration') is 'exception'
PASS TryAllocate('CSSStyleRule') is 'exception'
PASS TryAllocate('DOMImplementation') is 'exception'
PASS TryAllocate('HTMLCollection') is 'exception'
PASS TryAllocate('MediaList') is 'exception'
PASS TryAllocate('MimeType') is 'exception'
PASS TryAllocate('MimeTypeArray') is 'exception'
PASS TryAllocate('MutationEvent') is 'exception'
PASS TryAllocate('NamedNodeMap') is 'exception'
PASS TryAllocate('NodeFilter') is 'exception'
PASS TryAllocate('NodeList') is 'exception'
PASS TryAllocate('Plugin') is 'exception'
PASS TryAllocate('PluginArray') is 'exception'
PASS TryAllocate('StyleSheet') is 'exception'
PASS TryAllocate('StyleSheetList') is 'exception'
PASS TryAllocate('TextEvent') is 'exception'
PASS TryAllocate('XPathResult') is 'exception'
PASS TryAllocate('BarProp') is 'exception'
PASS TryAllocate('CanvasGradient') is 'exception'
PASS TryAllocate('CanvasPattern') is 'exception'
PASS TryAllocate('Selection') is 'exception'
PASS TryAllocate('Window') is 'exception'
PASS TryAllocate('History') is 'exception'
PASS TryAllocate('HTMLOptionsCollection') is 'exception'
PASS TryAllocate('Location') is 'exception'
PASS TryAllocate('Navigator') is 'exception'
PASS TryAllocate('NodeIterator') is 'exception'
PASS TryAllocate('Screen') is 'exception'
PASS TryAllocate('TreeWalker') is 'exception'
PASS TryAllocate('XPathExpression') is 'exception'
PASS TryAllocate('Worker') is 'exception'
PASS TryAllocate('EventTargetNode') is 'no constructor'
PASS TryAllocate('UndetectableHTMLCollection') is 'no constructor'
PASS TryAllocate('XPathNSResolver') is 'no constructor'
PASS TryAllocate('EventListener') is 'no constructor'
PASS TryAllocate('NPObject') is 'no constructor'
PASS TryAllocate('Comment') is '[object Comment]'
PASS TryAllocate('Comment') is '[object Comment]'
PASS TryAllocate('Comment') is '[object Comment]'
PASS TryAllocate('CSSStyleSheet') is '[object CSSStyleSheet]'
PASS TryAllocate('CSSStyleSheet') is '[object CSSStyleSheet]'
PASS TryAllocate('CSSStyleSheet') is '[object CSSStyleSheet]'
PASS TryAllocate('DataTransfer') is '[object DataTransfer]'
PASS TryAllocate('DataTransfer') is '[object DataTransfer]'
PASS TryAllocate('DataTransfer') is '[object DataTransfer]'
PASS TryAllocate('DOMParser') is '[object DOMParser]'
PASS TryAllocate('DOMParser') is '[object DOMParser]'
PASS TryAllocate('DOMParser') is '[object DOMParser]'
PASS TryAllocate('Document') is '[object Document]'
PASS TryAllocate('Document') is '[object Document]'
PASS TryAllocate('Document') is '[object Document]'
PASS TryAllocate('DocumentFragment') is '[object DocumentFragment]'
PASS TryAllocate('DocumentFragment') is '[object DocumentFragment]'
PASS TryAllocate('DocumentFragment') is '[object DocumentFragment]'
PASS TryAllocate('Range') is '[object Range]'
PASS TryAllocate('Range') is '[object Range]'
PASS TryAllocate('Range') is '[object Range]'
PASS TryAllocate('Text') is '[object Text]'
PASS TryAllocate('Text') is '[object Text]'
PASS TryAllocate('Text') is '[object Text]'
PASS TryAllocate('XMLHttpRequest') is '[object XMLHttpRequest]'
PASS TryAllocate('XMLHttpRequest') is '[object XMLHttpRequest]'
PASS TryAllocate('XMLHttpRequest') is '[object XMLHttpRequest]'
PASS TryAllocate('XMLSerializer') is '[object XMLSerializer]'
PASS TryAllocate('XMLSerializer') is '[object XMLSerializer]'
PASS TryAllocate('XMLSerializer') is '[object XMLSerializer]'
PASS TryAllocate('XPathEvaluator') is '[object XPathEvaluator]'
PASS TryAllocate('XPathEvaluator') is '[object XPathEvaluator]'
PASS TryAllocate('XPathEvaluator') is '[object XPathEvaluator]'
PASS TryAllocate('XSLTProcessor') is '[object XSLTProcessor]'
PASS TryAllocate('XSLTProcessor') is '[object XSLTProcessor]'
PASS TryAllocate('XSLTProcessor') is '[object XSLTProcessor]'
PASS TryAllocate('Audio') is '[object HTMLAudioElement]'
PASS TryAllocate('Audio') is '[object HTMLAudioElement]'
PASS TryAllocate('Audio') is '[object HTMLAudioElement]'
PASS TryAllocate('Option') is '[object HTMLOptionElement]'
PASS TryAllocate('Option') is '[object HTMLOptionElement]'
PASS TryAllocate('Option') is '[object HTMLOptionElement]'
PASS TryAllocate('Image') is '[object HTMLImageElement]'
PASS TryAllocate('Image') is '[object HTMLImageElement]'
PASS TryAllocate('Image') is '[object HTMLImageElement]'
PASS successfullyParsed is true

TEST COMPLETE
